requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cloudinary==1.42.1
colorama==0.4.6
dnspython==2.7.0
fastapi==0.115.7
gevent==24.11.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
pillow==11.1.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pymongo==4.11
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
websocket==0.2.1
zope.event==5.0
zope.interface==7.2
